refactor: split OperationalDiagnostics.swift (1,464 lines → 4 files)
Largest file in the repo, per #491.  The monolithic class held three
distinct concerns: event recording, snapshot/metrics aggregation, and
DB query/projection helpers.  Each is now its own extension file:

  OperationalDiagnostics.swift            — types, class skeleton,
                                            workerModeTestSetupIDs,
                                            free helpers + storage (281)
  OperationalDiagnostics+Recording.swift  — recordSubmissionCreated,
                                            recordRunnerCheckIn,
                                            recordJobAssigned,
                                            recordWorkerResult, … (596)
  OperationalDiagnostics+Metrics.swift    — rollingAverages,
                                            metricsSnapshot,
                                            timeseries, prune,
                                            recordRequestMetric (231)
  OperationalDiagnostics+Queries.swift    — findOrCreate*, queue depth,
                                            peakLoad, durationSummary,
                                            *Metadata builders (396)

Visibility changes (internal-only, no external API impact):
- `maintenance` + `compatibilityCounters` properties dropped `private`
  so extension methods in sibling files can access them.
- File-private free helpers `millisecondsBetween` + `iso8601Metadata`
  dropped `private` (called from Recording).
- `SubmissionDiagnosticsContext` dropped `private` (used by both
  Recording and Queries extensions).
- The original `private extension OperationalDiagnosticsService` for
  query helpers becomes a plain `extension` so Recording can call into
  it across the file boundary.

Build + test pass.
